% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{multiply_by_two_copy}
\alias{multiply_by_two_copy}
\title{Returns a copy with each element multiplied by two.}
\usage{
multiply_by_two_copy(integers)
}
\description{
Returns a copy with each element multiplied by two.
}
