% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{multiply_copy}
\alias{multiply_copy}
\title{Returns a copy with each element multiplied by two.}
\usage{
multiply_copy(doubles)
}
\description{
Returns a copy with each element multiplied by two.
}
